+++ maven/plugin/xdb-maven-plugin/src/test/java/org/safris/maven/plugin/xdb/xde/GroupClauseTest.java
@@ -42,6 +42,6 @@ public class GroupClauseTest extends LoggableTest {
       execute();
 
     Assert.assertTrue(rows.nextRow());
-    Assert.assertEquals(Long.valueOf(1), rows.nextEntity().get());
+    Assert.assertEquals(Integer.valueOf(1), rows.nextEntity().get());
   }
 }
